@@ -1,9 +1,9 @@
 [project]
 name = "bazel-pyproject"
 version = "0"
-requires-python = ">=3.9" # NOTE: Must be kept in sync with other global settings
+requires-python = ">=3.10" # NOTE: Must be kept in sync with other global settings
 dependencies = [
-  "editdistance",
+  "levenshtein",
   "absl-py",
   "aiofiles",
   "aiohttp",
