@@ -1,8 +1,10 @@
 [project]
 name = "modular"
 version = "0"
-requires-python = ">= 3.9"
+requires-python = ">= 3.10"
 
+
+# TODO: restrict utils/packaging/conda/test_inputs to only test_format.mojo
 [tool.black]
 include = '\.mojo$'
 line-length = 80
