@@ -11,11 +11,12 @@
 # limitations under the License.
 # ===----------------------------------------------------------------------=== #
 
-# DOC: max/tutorials/build-an-mlp-block.mdx
+# DOC: max/develop/build-an-mlp-block.mdx
 
 from __future__ import annotations
 
-from typing import Any, Callable, Optional
+from collections.abc import Callable
+from typing import Any
 
 from max import nn
 from max.dtype import DType
@@ -27,8 +28,8 @@ class MLPBlock(nn.Module):
         self,
         in_features: int,
         out_features: int,
-        hidden_features: Optional[list[int]] = None,
-        activation: Optional[Callable[[TensorValue], TensorValue]] = None,
+        hidden_features: list[int] | None = None,
+        activation: Callable[[TensorValue], TensorValue] | None = None,
     ) -> None:
         super().__init__()
 
