@@ -12,10 +12,10 @@
 # limitations under the License.
 # ===----------------------------------------------------------------------=== #
 
+import os
 from collections.abc import Sequence
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Optional
 
 import click
 import numpy as np
@@ -42,7 +42,7 @@ MARKER_COLORS = [
 ]
 
 
-def _get_max_groups_per_chart(x_labels, width_px, font_size_pt):  # noqa: ANN001
+def _get_max_groups_per_chart(x_labels, width_px, font_size_pt):  # noqa: ANN001, ANN202
     x_labels_len = [
         ([len(t) for t in entry.split("<br>")]) for entry in x_labels
     ]
@@ -190,7 +190,7 @@ def draw_plot(
         )
 
 
-def get_labels_by_pivots(x_labels, pivots):  # noqa: ANN001
+def get_labels_by_pivots(x_labels, pivots):  # noqa: ANN001, ANN201
     df = label_to_df(x_labels)
     assert len(df) == len(x_labels)
     pivot_labels = []
@@ -202,11 +202,11 @@ def get_labels_by_pivots(x_labels, pivots):  # noqa: ANN001
     return pivot_labels
 
 
-def wrap_labels(x_labels):  # noqa: ANN001
+def wrap_labels(x_labels):  # noqa: ANN001, ANN201
     return [entry.replace("$", "") for entry in x_labels]
 
 
-def label_to_df(x_labels):  # noqa: ANN001
+def label_to_df(x_labels):  # noqa: ANN001, ANN201
     ds = []
     for label in x_labels:
         vals = label.replace("$", "").split("/")
@@ -220,7 +220,7 @@ def label_to_df(x_labels):  # noqa: ANN001
     return df
 
 
-def extract_pivots(x_labels):  # noqa: ANN001
+def extract_pivots(x_labels):  # noqa: ANN001, ANN201
     df = label_to_df(x_labels)
     pivot_columns = []
     for c in df.columns:
@@ -232,7 +232,7 @@ def extract_pivots(x_labels):  # noqa: ANN001
     return pivot_columns, non_pivot_columns
 
 
-def append_wrap_fixed_width(lst: list[str], sep: str, num_lines: int = 2):
+def append_wrap_fixed_width(lst: list[str], sep: str, num_lines: int = 2):  # noqa: ANN201
     s: list[str] = []
     result = []
     current_len = 0
@@ -467,7 +467,7 @@ def parse_and_plot(
 def cli(
     input_files,  # noqa: ANN001
     label_list: list[str],
-    output_prefix: Optional[str],
+    output_prefix: str | None,
     plot_col: str,
     key: str,
     compare: bool,
@@ -525,4 +525,6 @@ def main() -> None:
 
 
 if __name__ == "__main__":
+    if directory := os.environ.get("BUILD_WORKING_DIRECTORY"):
+        os.chdir(directory)
     main()
