@@ -22,7 +22,6 @@ import sys
 import warnings
 from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Optional
 
 import click
 import numpy as np
@@ -39,8 +38,7 @@ warnings.filterwarnings("ignore")
 LINE = 80 * "-"
 
 
-HEADER = """
-# ===----------------------------------------------------------------------=== #
+HEADER = """##===----------------------------------------------------------------------===##
 # Copyright (c) 2025, Modular Inc. All rights reserved.
 #
 # Licensed under the Apache License v2.0 with LLVM Exceptions:
@@ -51,7 +49,7 @@ HEADER = """
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-# ===----------------------------------------------------------------------=== #
+##===----------------------------------------------------------------------===##
 """
 
 
@@ -77,7 +75,7 @@ def specs_to_df(specs: list[str]) -> pd.DataFrame:
     return df
 
 
-def extract_pivots_df(df: pd.DataFrame, exclude: Optional[list[str]] = None):
+def extract_pivots_df(df: pd.DataFrame, exclude: list[str] | None = None):  # noqa: ANN201
     if exclude is None:
         exclude = []
     # df = specs_to_df(x_labels)
@@ -96,12 +94,12 @@ def extract_pivots_df(df: pd.DataFrame, exclude: Optional[list[str]] = None):
     return pivot_columns, non_pivot_columns
 
 
-def extract_pivots(x_labels: list[str], exclude: Optional[list[str]] = None):
+def extract_pivots(x_labels: list[str], exclude: list[str] | None = None):  # noqa: ANN201
     df = specs_to_df(x_labels)
     return extract_pivots_df(df=df, exclude=exclude)
 
 
-def load_pickle(path):  # noqa: ANN001
+def load_pickle(path):  # noqa: ANN001, ANN201
     with open(path, "rb") as handle:
         return pickle.load(handle)
 
@@ -114,7 +112,7 @@ def dump_yaml(obj, out_path) -> None:  # noqa: ANN001
     yaml.dump(obj, sys.stdout, sort_keys=False)
 
 
-def top_idx(x, top_percentage=0.05):  # noqa: ANN001
+def top_idx(x, top_percentage=0.05):  # noqa: ANN001, ANN201
     # calculate the threshold to pick top_percentage of the results
     threshold = (top_percentage + (np.min(x) / np.max(x))) * np.max(x)
     return x.where(x < threshold).dropna().index
@@ -132,7 +130,7 @@ def replace_vals_snippet(p_spec, snippet_path) -> str:  # noqa: ANN001
     return c
 
 
-def find_common_params(subset):  # noqa: ANN001
+def find_common_params(subset):  # noqa: ANN001, ANN201
     spec_list = []
     for index, row in subset.iterrows():  # noqa: B007
         p = spec_to_dict(row["spec"])
@@ -164,7 +162,7 @@ def df_to_console_table(
         style = col_style.get(c, None)
         table.add_column(c, justify="left", style=style)
 
-    def wrap(x):  # noqa: ANN001
+    def wrap(x):  # noqa: ANN001, ANN202
         return "\n".join(x.split("/"))
 
     for row in df.itertuples(index=index):
@@ -181,7 +179,7 @@ class TuningSpec:
     name: str = ""
     file: Path = Path("")
     params: list[dict] = field(default_factory=list)
-    pkl_path: Path = Path()
+    src_path: Path = Path()
     git_sha: str = ""
     datetime: str = ""
 
@@ -238,7 +236,7 @@ class KbenchPKL:
         return f
 
 
-def df_round_floats(df, prec=3):  # noqa: ANN001
+def df_round_floats(df, prec=3):  # noqa: ANN001, ANN201
     "Round values in dataframe to specified precision"
     for c in df.columns:
         if df.dtypes[c] in (np.float64, np.float32):
@@ -255,7 +253,7 @@ def profile_results(
     metric: str = "met (ms)",
     pivots: list[str] = [],  # noqa: B006
     verbose=False,  # noqa: ANN001
-) -> Optional[TuningSpec]:
+) -> TuningSpec | None:
     try:
         pkl = KbenchPKL(pickle_path=pickle_path, metric=metric)
     except:
@@ -324,7 +322,7 @@ def profile_results(
         name=str(pkl_data.get("name", None)),
         file=Path(pkl_data.get("file", Path())),
         params=[spec],
-        pkl_path=pickle_path,
+        src_path=pickle_path,
         git_sha=str(pkl_data.get("git-revision", None)),
         datetime=str(pkl_data.get("datetime", None)),
     )
@@ -351,7 +349,21 @@ def yaml_to_tuning_spec(yaml_path: Path) -> list[TuningSpec]:
             name=s.name,
             file=s.file,
             params=[spec_to_dict(str(s))],
-            pkl_path=yaml_path,
+            src_path=yaml_path,
+            git_sha="",
+        )
+        tuning_specs.append(ts)
+    return tuning_specs
+
+
+def spec_to_tuning_spec(spec: Spec) -> list[TuningSpec]:
+    tuning_specs = []
+    for s in spec:
+        ts = TuningSpec(
+            name=s.name,
+            file=s.file,
+            params=[spec_to_dict(str(s))],
+            src_path=Path(),
             git_sha="",
         )
         tuning_specs.append(ts)
@@ -376,10 +388,6 @@ def diff_baseline(
 
     # Find the common pivots between all pkl's if none specified
     if not pivots:
-
-        def intersection(a: list, b: list):
-            return [x for x in a if x in b]
-
         _, pivots = extract_pivots(
             list(tune_df_base["spec"]), exclude=["name", "AUTOTUNING_MODE"]
         )
@@ -388,7 +396,7 @@ def diff_baseline(
             _, non_pivot_columns = extract_pivots(
                 list(pkl.tune_df["spec"]), exclude=["name", "AUTOTUNING_MODE"]
             )
-            pivots = intersection(pivots, non_pivot_columns)
+            pivots = list_intersection(pivots, non_pivot_columns)
 
     shape = "/".join([f"{p}={base_dict[p]}" for p in pivots])
     for i, f in enumerate(files[1:]):
@@ -438,7 +446,7 @@ def codegen_snippet(
     for idx, s in enumerate(specs):
         config_str = replace_vals_snippet(s.params[0], snippet_path)
         print(LINE)
-        details += [f"# Automatically generated from [{s.pkl_path}]"]
+        details += [f"# Automatically generated from [{s.src_path}]"]
         details += [f"# index: [{idx}]"]
         if s.datetime:
             details += [f"# date: [{s.datetime}]"]
@@ -467,7 +475,7 @@ def codegen_snippet(
     print(f"wrote results to [{abs_output_path}]")
 
 
-def yaml_reference_handling(s: str):
+def yaml_reference_handling(s: str):  # noqa: ANN201
     # all of them should be uniq
     ref_pattern = re.compile(r"('<<'[\s]*:[\s]*'*)([^']*)'")
     refs = re.findall(ref_pattern, s)
@@ -494,8 +502,7 @@ def codegen_yaml(specs: list[TuningSpec], output_path: Path) -> None:
     # collect first set of parameters (i.e., "params[0]") from all specs.
     params = [s.params[0] for s in specs]
     df = pd.DataFrame(params)
-    _, non_pivots = extract_pivots_df(df, exclude=[])
-    print("common columns", non_pivots)
+    pivots, non_pivots = extract_pivots_df(df, exclude=[])
 
     # remove common columns (non-pivots) to get unique rows of data
     uniq_rows = df.drop(columns=non_pivots)
@@ -527,13 +534,16 @@ def codegen_yaml(specs: list[TuningSpec], output_path: Path) -> None:
     with open(output_yaml, "w") as f:
         f.write(yaml_str)
 
+    print(f"common columns: {non_pivots}")
+    print(f"pivot columns: {pivots}")
+    print(f"num entries: [{len(uniq_rows)}]")
     print(f"wrote results to [{output_yaml}]")
     print(LINE)
 
 
 # TODO: add more checks for inconsistency between various input files.
 def check_specs(
-    specs: list[TuningSpec], key_cols: Optional[list[str]] = None
+    specs: list[TuningSpec], key_cols: list[str] | None = None
 ) -> bool:
     # TODO: check specs have the same tuning hash
     spec_list = [pd.DataFrame([s.params[0]]) for s in specs]
@@ -559,13 +569,74 @@ def check_specs(
         )
 
 
+def list_intersection(a: list, b: list) -> list:
+    return [x for x in a if x in b]
+
+
+def merge_specs(
+    spec_list: list[Spec],
+    pivots: list[str] | None = None,
+    sort_pivots: list[str] | None = None,
+) -> pd.DataFrame:
+    """
+    Merges a list of Spec objects into a single pandas DataFrame, join on specified pivot columns.
+
+    Args:
+        spec_list (list[Spec]): List of Spec objects to merge.
+        pivots (list[str] | None, optional): List of column names to use as pivots for merging. If None, uses the intersection of all columns in the DataFrames.
+        sort_pivots (list[str] | None, optional): List of column names to sort the merged DataFrame by. If None, uses the pivots.
+
+    Returns:
+        pd.DataFrame: Merged DataFrame containing the specified pivots and a 'spec' column with the corresponding Spec objects.
+    """
+    df_list = []
+    for spec in spec_list:
+        df_list += [specs_to_df([str(s) for s in spec])]
+
+    # limit to a subset of pivots, if not select them all
+    if not pivots:
+        pivots = list(df_list[0].columns)
+        for df in df_list[1:]:
+            pivots = list_intersection(pivots, list(df.columns))
+
+    # assert all pivots are present in columns of all df's
+    for i, df in enumerate(df_list):
+        assert np.all([p in df.columns for p in pivots])
+        # ignore non-pivot columns
+        df = df[pivots]
+        df["spec"] = spec_list[i]
+        df_list[i] = df
+
+    # merge all df's
+    df_merged = df_list[0]
+    for df in df_list[1:]:
+        df_merged = pd.merge(df_merged, df, on=pivots, how="outer")
+
+    # Fill 'spec' with 'spec_x' if 'spec_x' is not NaN, otherwise with 'spec_y'.
+    # Then, drop 'spec_x' and 'spec_y'.
+    df_merged["spec"] = np.where(
+        df_merged["spec_x"].notna(), df_merged["spec_x"], df_merged["spec_y"]
+    )
+    df_merged = df_merged.drop(["spec_x", "spec_y"], axis=1)
+
+    if not sort_pivots:
+        sort_pivots = pivots
+    assert set(sort_pivots).issubset(set(pivots))
+    df_merged = (
+        df_merged.drop_duplicates(sort_pivots)
+        .sort_values(by=sort_pivots)
+        .reset_index(drop=True)
+    )
+    return df_merged
+
+
 class ComplexParamList(click.Option):
     """Complex parameter list
     Example:
         --pivot=[M] --pivot=[N] --pivot=[K] is equivalent to --pivot=[M,N,K] and vice versa.
     """
 
-    def type_cast_value(self, ctx, value_in):  # noqa: ANN001
+    def type_cast_value(self, ctx, value_in):  # noqa: ANN001, ANN201
         """DO NOT REMOVE this function, it is called from ctx in click."""
         p = []
         assert isinstance(value_in, list)
@@ -734,6 +805,24 @@ help_str = "Profile kbench output pickle"
     multiple=True,
 )
 @click.option(
+    "--sort-pivots",
+    cls=ComplexParamList,
+    default=[],
+    help="Specify the pivots to sort values.",
+    multiple=True,
+)
+@click.option(
+    "--merge",
+    is_flag=True,
+    default=False,
+    help="Merge the first incoming pkl/yamls with the first one.",
+)
+@click.option(
+    "--query",
+    default=None,
+    help="Query the yaml. Example: kprofile --merge *yaml --query 'M>64 and K<1024' ",
+)
+@click.option(
     "--correlation",
     "-c",
     is_flag=True,
@@ -759,7 +848,10 @@ def cli(
     diff,  # noqa: ANN001
     metric,  # noqa: ANN001
     pivots,  # noqa: ANN001
-    correlation,  # noqa: ANN001
+    sort_pivots,  # noqa: ANN001
+    merge: bool,
+    query: str,
+    correlation: bool,
     verbose,  # noqa: ANN001
 ) -> bool:
     if not verbose:
@@ -777,7 +869,32 @@ def cli(
 
     top_percentage = float(top) if top else 0
 
-    if diff:
+    # TODO: refactor profile, diff, merge, correlation into separate commands
+
+    if merge:
+        # All files should be .yaml
+        assert np.all([Path(path).suffix == ".yaml" for path in files])
+        spec_list = [Spec.load_yaml(Path(path)) for path in files]
+        merged_df = merge_specs(
+            spec_list, pivots=pivots, sort_pivots=sort_pivots
+        )
+
+        if query:
+            print(f"Applying query: [{query}]")
+            merged_df = merged_df.query(query).reset_index(drop=True)
+        if verbose:
+            print(merged_df.loc[:, merged_df.columns != "spec"].to_string())
+            print(LINE)
+        if len(merged_df) == 0:
+            raise ValueError(
+                "Query resulted in empty dataset. No data matches the specified query or filters."
+            )
+
+        specs = spec_to_tuning_spec(merged_df["spec"])
+
+        codegen_yaml(specs, output_path=output_path)
+
+    elif diff:
         if head == -1:
             head = 1
         diff_baseline(
@@ -786,7 +903,6 @@ def cli(
     else:
         specs = []
         invalid_pkls = []
-        # for pkl_path in files:
         for path in files:
             suffix = Path(path).suffix
             if suffix == ".yaml":
@@ -842,4 +958,6 @@ def main() -> None:
 
 
 if __name__ == "__main__":
+    if directory := os.environ.get("BUILD_WORKING_DIRECTORY"):
+        os.chdir(directory)
     main()
