@@ -11,7 +11,7 @@
 # limitations under the License.
 # ===----------------------------------------------------------------------=== #
 
-# DOC: max/tutorials/get-started-with-max-graph-in-python.md
+# DOC: max/develop/get-started-with-max-graph-in-python.mdx
 
 import numpy as np
 from max import engine
