@@ -20,7 +20,7 @@ try:
     from IPython.core.magic import register_cell_magic  # type: ignore
 except ImportError:
 
-    def register_cell_magic(fn):  # noqa: ANN001
+    def register_cell_magic(fn):  # noqa: ANN001, ANN201
         return fn
 
 
@@ -79,7 +79,7 @@ def mojo(line, cell) -> None:  # noqa: ANN001
             %%mojo package -o kernels.mojopkg
 
             from runtime.asyncrt import DeviceContextPtr
-            from tensor_internal import InputTensor, ManagedTensorSlice, OutputTensor
+            from tensor import InputTensor, ManagedTensorSlice, OutputTensor
 
             @compiler.register("histogram")
             struct Histogram:
